fix: remove return_fileds from hybrid_search
